+++ build_runner_core/CHANGELOG.md
@@ -8,6 +8,9 @@
 
 ### Other changes
 
+- Builds no longer depend on the contents of the package_config.json file,
+  instead they depend only on the language versions inside of it.
+  - This should help CI builds that want to share a cache across runs.
 - Remove unused dev dependency on `package_resolver`.
 
 ## 4.5.3